Bug 1400222 - Always pass an explicit ForeignKey on_delete value
The default value for on_delete is `CASCADE`, so this change is a
no-op aside from fixing hundreds of deprecation warnings of form:

```
RemovedInDjango20Warning: on_delete will be a required arg for
ForeignKey in Django 2.0. Set it to models.CASCADE on models and in
existing migrations if you want to maintain the current default
behavior.
```
(Plus the same for `OneToOneField`)

See:
https://docs.djangoproject.com/en/1.11/ref/models/fields/#django.db.models.ForeignKey.on_delete