Bug 1416257 - Disable yarn strict engines mode using .yarnrc
As of yarn 1.0+ if the version of node or yarn doesn't match that
specified in `package.json`, yarn commands will fail to run and
complain about incompatible versions. Unfortunately the error message
doesn't really make it clear that the user has the wrong version
installed, and instead gives the impression something is broken with
Treeherder. In addition, much of the time the exact version isn't
important, so the user doesn't need to go to the trouble of changing
their node installation.

As such, we just disable the yarn engines check for now.