Bug 1421073 - Vagrant: Fix networking on non-Windows platforms (#3015)
On platforms that are not Windows, the Vagrantfile enables a second
virtual network adapter to allow NFS to work. However this breaks
the fix in bug 1417099, since it only expected one interface to be
returned by the grepped `ip` output. We could have adjusted that fix
to instead iterate through the results, however it's easier to just
use a wildcard with `iptables` instead, since we know all ethernet
adapters start with the letter `e`.