Bug 1432840 - Fix/enable eslint 'import/no-named-as-default'
In `configureStore.js` the same object was being exported twice, once
as the default export and once as a named export. Since default exports
are preferred if there is only one export in a file, I've removed the
named import and left the default one.

In `Groups.jsx` the `Groups` class was exported but unused, so has
been adjusted to no longer be exported, so the `App.jsx` import
doesn't trigger the warning:
`import Groups from './Groups';`

See:
https://github.com/benmosher/eslint-plugin-import/blob/master/docs/rules/no-named-as-default.md