Remove usage of taskcluster-client
Previously, taskcluster-client-web was used to get taskcluster
credentials but taskcluster-client was then used to setup the client. I
believe this was causing some issues with expiration since the agent
creates credentials that are only good for a couple of minutes.

To fix this issue, this commit uses taskcluster-client-web to handle
everything. This commit brings other benefits, namely:
- taskcluster-client is no longer a dependency.
- thTaskcluster is now written in plain JS. This should drive the
transition from angular to react in the right direction.