Bug 1442556 - Improve rendering speed in treeherder by removing refs (#3307)
These refs were not actually used, so can be removed.  It does appear to
improve rendering speed, especially when changin filtering.  I had noticed
that the JobGroup was re-rendering more times than it should, and this
change stops that.  I believe more work is to be done here, but this is one
step.

This also includes a couple other small tweaks that may help speed.