Bug 1451342 - Unvendor ng-text-truncate-2
Since we've moving away from vendored packages.

The only differences between the vendored file and that in the NPM
package were the changes to `require('angular')` and export the
created AngularJS module. However Angular is already imported prior
to this package being imported, and we can just pass the module name
as a string instead of relying on the export.