Bug 1443129 - Add prop types to all JSX components (#3379)
* enable react/prop-types eslint rule
While enabling this, I set all the PropTypes as ``isRequired`` then ran the app
and wherever I hit issues where the value was undefined, then I removed ``isRequired``.

* Enable react/require-default-props eslint rule
In some cases "null" was the best default.  I think this is OK.  But having empty arrays
for others meant less checking in the actual code, so bonus!  :)