Bug 1451774 - Remove bytestring attributes from existing migrations
Now that the models / `settings.py` contain the `unicode_literal`
import, we either need to generate a new migration to bring the
migrations in sync with the model, or else manually edit the existing
migrations.

We have to do the latter for the `replaces` property instances due to:
https://code.djangoproject.com/ticket/29290

...so for consistency (and to save creating a pointless migration),
we do so for the other instances too.