Bug 1463079 - Remove the preAssignBindingsEnabled workaround
AngularJS 1.6 changed the way controller bindings were initialised,
but allowed users to revert to the previous behaviour by setting
`preAssignBindingsEnabled(true)`.

As part of our upgrade from AngularJS 1.5 to 1.6, we set that pref
in Perfherder, to fix the bug 1428958 regression.

However AngularJS 1.7 has now removed the pref, so we need to actually
fix the root cause - by moving any controller logic that is dependant
on the bindings into the `.$onInit()` method, which is only called
after all of the bindings are guaranteed to exist.

See:
https://docs.angularjs.org/guide/migration#migrate1.5to1.6-ng-services-$compile

Testing this locally shows the regression in bug 1428958 no longer
occurs, and as far as I can tell all other functionality in Perfherder
works fine (and there are no new exceptions seen in the console).