Bug 1465041 - Restore default webpack module resolution config
Neutrino 4 removes the webpack default module resolution of looking
in a relative `node_modules` directory (and then the parent of that
directory, and so on), in favour of hardcoding the repo root's
`node_modules` and those of the Neutrino preset directories.

This is both unnecessary and causes incorrect module resolution,
which has bitten us multiple times.