Filter errors-we-care-about down to those with attached FailureLines
A Job has N related TextLogErrors and a subset of those have a related
FailureLine.  When matching we only care about the subset with
FailureLines.  Instead of calculating that subset from the mutated pool
attached to the Job we calculate it up front but maintain the full list.
This lays the groundwork for not mutating the full list later.