Remove ENABLE_DEBUG_TOOLBAR setting
This was added because the debug toolbar dependency is defined in
requirements/dev.txt and it allowed enabling DEBUG in a deployment
environment (like prototype) where the toolbar might not be deployed.
However we now feel this is an unlikely (and unwanted) position to be in
(using DEBUG) so we're making DEBUG synonymous with the toolbar.