Bug 1470532 - Fix fixed by commit after intermittent
Ends up I had some cruft in there from the Angular days.  Saving a
``classification`` in the state was wrong.  We store the values for
it (id and comment) in the state, then create a ``classification``
object based on those values for ``POST``ing back to the API during
``save``.