separate TextLogError custom router from default routing
DRF treats "detail" views as those with a PK/ID in them and only
provides a PUT handler for URLs that match that format.  Our
TextLogErrorViewSet needs to handle PUTs made to non-detail-style URLs
so we can PUT changes to more than one item at a time.  However since no
other ViewSet needs this functionality this commit pulls it out into a
special case Router so it's clear why it exists.