Remove settings for Durable and Auto Delete of Pulse queues
Since these settings aren't configurable from the env and are buried
deep in the global settings they don't add anything to the
configurability of the project.  It's highly likely that anyone changing
these settings for local development has deep knowledge of Kombu/Pulse
and can just as easily change the code.