Replace BugJobMap.objects.create() with BugJobMap.create()
Whenever we create a BugJobMap with a User we want to update its Job's
best FailureClassification's Bug number with the one the BugJobMap was
created with.  This keeps the autoclassify data (FailureClassification)
in sync with BugJobMap.

However, doing this via an overridden `.save()` effectively hides the
functionality from anyone reading through the code.  Since the
`.update_autoclassification_bug()` method is only called from that one
place this moves all the functionality into the classmethod `.create()`.
This makes it explicit that creating a BugJobMap involves more than a
simple DB row creation.