Move Pulse Publishing to Pulse Service (#3919)
The PulsePublisher class was built for extensibility, providing lots of hooks
for customisation.  However we only had one subclass in use since its
introduction: TreeherderPublisher.  This reduces the concrete class into a
single function which publishes the given message.  In doing so all
configurability has been removed, since it was unused.