Bug 1450040 - Rename usage internal to PushJobs from "platforms" to "filteredPlatforms"
This also simplifies getting a job by id.  No longer requires the "jobMap" and just
uses the React DOM instance.  This isn't a perfect diff because it pulls a couple
props imports that will come in with the final commit, but hopefully you get the gist
of this to remove noise from that bigger commit.  :)