Bug 1499551 - Address memory leak when loading new pushes
This fix is not a "smoking gun" kind of fix.  But it does appear to remedy
the problem in local testing.  It is uses the spread syntax rather than
adding to the existing push, which I believe to be more correct.

It's my belief that the pushList.push was what was causing problems.