Bug 1501984 - Fix use of class-level variables (#4445)
For PushActionMenu, we were using class variables to cache props.
I'm not sure why I did that back in February, but this corrects it.

For Pushes context, this stores the cache of cachedReloadTriggerParams
in a class variable instead of the state.  State changes should be for
changes that affect what's displayed in the UI.  But this variable is not
shown in the UI.  It's used to determine whether or not to reload the
page based on certain URL params changing.  So this change saves having
to add a shouldComponentUpdate check where it's not otherwise needed.