Bug 1364045 - Convert all unit tests to Jest
Also remove all Karma support and update the docs to only mention ``Jest``.

One of the test files was testing some AngularJS filters. I converted these
tests to test the equivalent helper functions.

It was a little tricky getting ``fetch`` to work in all the tests.  These tests are
mocking ``fetch`` calls with ``fetch-mock``.  There is a ``jest-fetch-mock`` but it
does not look as good.  You have to specify the fetch responses in order, rather
than mapped by the request itself.  So I decided to stick with ``fetch-mock``.
I filed an upstream issue with ``fetch-mock``:
https://github.com/wheresrhys/fetch-mock/issues/413