Vagrant: Fix error during dist-upgrade due to conf file prompt
Recent Ubuntu package updates mean that a clean Vagrant provision was
failing due to dpkg wanting an answer to the "which version of the config
file would you like to keep" question, which fails when apt-get is run
non-interactively.

To avoid such prompts we have to force dpkg to always pick one of the
config file conflict resolution options -- in this case I've chosen to
always use the new version of the config file (via `--force-confnew`),
since we don't customise any system config files, so overwriting any
existing ones is fine.