Bug 1270157 - Make the CSP report collector skip CSRF checks
Since otherwise report submissions will fail with an HTTP 403 on browsers
that send the Django session cookie but not the CSRF token.

The collector has also been rewritten to be a standard Django view rather
than a django-rest-framework APIView, since the latter is more of a
hindrance than a help for this use-case (particularly now we're wanting
to disable CSRF checks).