Bug 1530206 - Remove the now emptied store_pulse_resultsets queue
In #4709 push ingestion was changed to use a new `store_pulse_pushes`
queue (since we now prefer the term "push" instead of "resultset"), with
the old queue left behind to ensure any tasks in it at the time of
deployment would still be run.

Now that the old queue is empty it can be removed.