Bug 1539580 - platform 'opt' should be on top of list for operating system
The 'opt' build flavor is supposed to be shown on top for the group of platforms whose only difference is the build flavor:
https://github.com/mozilla/treeherder/blob/fa39de7925dc2dc398c22389d31fa7af79369ab5/ui/helpers/constants.js#L125

Because its value is 0, it evals to false in the if check at
https://github.com/mozilla/treeherder/blob/fa39de7925dc2dc398c22389d31fa7af79369ab5/ui/job-view/pushes/Push.jsx#L233
and 10 gets used as value instead which puts it at bottom.

Let's increase the value for each flavor so all are positive and the sort function doesn't have to become more complex.