Bug 1521157 - Accurately show pass/fail for push
Calculate size of needInvestigation and fixedByCommit to show
whether or not there were any issues with this push.

For each Metric, just show the pass/fail rather than the arbitrary number.
We may decide on a number at some point, but it doesn't have a meaning
right now.

Also move "Tests" to the top, since it's the only one with real data
for now.

Plus a few formatting fixes.