Revert "Bug 1346567 - Make dedicated cyclers" (#4854)
This reverts commit fa533ff25a2b7654e42760e854f1a537eca002f1.

Ends up there's a bug with this change. The plan is for @ionutgoldan to create a new PR with just the migration and model change for PerformanceDatum which we can deploy. Then he can re-implement his code changes from this PR and we'll already be done with the lengthy migration to add the index.