Bug 1521157 - Stop using model_to_dict
I had to switch away from using ``model_to_dict`` to convert the jobs to be ``JSON`` serializable because that function will strip away the ``_id`` from fields like ``job_type``.  This was fine in the Push Health UI itself, but when I started using the shared code of retriggering, then the ``job`` objects didn't have matching fields with what was in Treeherder.