Cleanup selected job redux (#5119)
* Create function for detecting filter changes on URL

* Prevent models from caching the current repo name

This is necessary so that we can change repos without having to reload the entire page.
We will stop reloading the page on repo change in a later PR.

* Don't use Redux when not necessary for "selectedJob"

I was using Redux to pass the selectedJob value down a single level
in a few cases.  So just pass it instead of invoking the reducer.

* Two small tune-up tweaks.

The "withPushes" was totally unnecessary on PushJobs
And Push can be a PureComponent.