Bug 1563014 - Enable the Django debug toolbar within Docker
Once we [switched to using Docker instead of Vagrant](https://github.com/mozilla/treeherder/pull/4901) we lost
the ability of using the Django debug toolbar when loading an API. This is a very important tool for Django development.

In Vagrant, we could specify which internal IPs to use, however, with Docker we can't predefine an IP before starting the container.

The toolbar allow us to provide options and specificall a function to determine to show or not the toolbar (via the
`SHOW_TOOLBAR_CALLBACK` option).