Adjust workers for Heroku Review Apps
Some workers are set to 0 dynos so we only focus on ingesting pushes and tasks. Nevertheless,
the number of workers can be adjusted after the review app is created. The current settings
are to get us the minimum amount of functionality for the app and the minimum throughput
for the queues not to backlog.