Display revision comments in RevisionInformation.
Fixes Bug 1487761.

The revision comments were supposed to be displayed,
but we were looking at the wrong place to do so.
This patch fixes this, so revision comments are displayed
under the revision number + author.
The layout is changed to flex to do this properly.
Colors are tweaked a bit in order to pass WCAG standard.