Manual ingestion - Disable eager consumption and allow specifying Gecko decision task
I noticed that we can ingest all tasks for a push without having to parse all logs by not
enabling eager celery consumption. If the user would like to do so it is still possible.

You can now also specify `--gecko-decision-task` to specify the Gecko decision task
associated to a push/revision. In the future we can determine this for the user but for
now we can just require it to be inputted manually.

A sample command:

```
./manage.py ingest push --project mozilla-central --revision fc3002cb85dc4d2c44d8aec1246f0356097411e3 --ingest-all-tasks --gecko-decision-task YiQiigCNQl63Hm21DjxXlA
```