SETA related changes (#5851)
* Development Docker container to initialize data

`docker-compose up` will now initialize the data for the local database.
This will also include initializing SETA's data.

* Change SETA's logging level and preseed information

The preseed information was all for Buildbot
The logging level changes is in preparation for when production switching to INFO messages instead of WARNING

* Docker container to log INFO messages instead of DEBUG messages

`LOGGING_LEVEL` is now set for the Docker container and it sets the logging
to `INFO` instead of `DEBUG`. The Django app will still be initialize in
debug mode but the default logging will be informative messages.

You can still log debugging messages by setting `LOGGING_LEVEL=DEBUG` before
initializing the Docker container.