Solve race condition for test path filtering
During review time[1] we introduced a small bug in the test filtering feature.

Reading `jobList` at the begining of the function instead of just before calling `mapPushJobs` almost guaranteed to start with an empty list which makes the call useless since it needs a non-empty list to work.

I also changed the code to call `this.setState` with a callback to `mapPushJobs` to guarantee that `jobTypeNameToManifests` is set in the state and not get an empty object.

[1] https://github.com/mozilla/treeherder/pull/5794#discussion_r369780355