Make coverage reporting fluctuate less
If any step on the Travis run fails (like formatting checks or a test) then
we don't report the code coverage for it.

Unfortunately, this can be distracting when we see status checks
reporting a failure or a big change.

This change makes code coverage to report even if there are minor
failures.