Bug 1509181 - Fix ingestion of GitHub pushes (#6065)
Github based projects can list pushes in the UI incorrectly. This is caused by commits that have been ammended.

This code change switches to grab the push time from the `timestamp` field of the head commit from the Pulse event rather that Github's APIs. This fixes the problem of push sorting.

Note, that the field `timestamp` only exists in the Push event that the Pulse message contains. This field can be seen in the `events` API, however, that API contains all sorts of events and holds a maximum of 300 events.

This change also includes partial support for manual ingestion of Git pushes.