Fix pre-commit's isort
`isort` as part of the pre-commit set up ends up under `$HOME/.cache/pre-commit`.

Due to that, it was unable to determine the path to the `setup.cfg` and was
unable to differentiate third party libraries from first party libraries.

Specifiying where the path to the default virtualenv is fixes it.