Docker: Use joint Pulse listener & unique queue name (#6288)
This change changes the Docker set up to use the new Pulse listener command that listens to both tasks and pushes instead of having two different ones.

This also adds a unique queue name. When running `docker-compose up` the Pulse listener will use a queue name that is the same for every developer running the listener. This means that all developers would be consuming from the same queue and stealing each other's Pulse messages. This change creates a unique name for each developer.

Co-authored-by: Armen Zambrano <armenzg@users.noreply.github.com>