Bug 1632403 Graph View - Refresh results in add new test modal when changing dropdowns
Tests filtered by the search box are handled in filteredData which are shown only when the text box is not empty, otherwise relatedTests or seriesData are shown. updateFilterText() is invoked only when the text in search box changes and never after one of the dropdowns are changed.
This PR is adding a call to updateFilterText() at the end of processOptions() call to update the results according to the changed dropdowns.