Make poetry the default method for generating the docs (#6424)
This makes sure that serving the documentation via poetry does not regress
and we have a single way of generating them.

Setting the build backend in `pyproject.toml` makes `pip` to use `poetry` to install the dependencies. By doing so, readthedocs will be able to install the required dependencies to generate the docs (since they don't support poetry directly).