Travis/jest - Disable printing to console on JS tests
Depending on how we write our tests I've noticed that it is not possible
to use `isRequired` as part of the `propTypes` without the tests calling
some unimportant component with all the required props.

In other words, this removes proptypes warnings in the output of our tests
while not preventing noticing test failures.

You can still test errors when they happen.