Fix issue where the bottom of the panel was cut off from view
This patch changes the details panel's container to use padding
rather than margin. With margin, the size of the container was using the
outer size, and cutting off the bottom of the contents. With padding,
the contents are correctly sized for the space and nothing is cut off.

A more correct fix would be to remove the margin and padding on the
outer container, and then ensure that the contents of each details panel
could determine its own padding. This would ensure the scroll bar is all
the way to the right and didn't cover any content. However, this would
be a riskier change, and would need each panel tested.