Add a table-based view for the performance data
This includes the unit and whether lower or higher is better. I felt a
bit confused looking at the results, and not knowing what the individual
values meant. This change also presents everything in a table format
which should make it easier to scan and understand.

I ran these changes by some users of this view, and handled some of
their early feedback. This included removing a graph that wasn't very
helpful. There might be some opportunities in the future to add more
details to the columns to help in making navigating the data easier.
However, this is not a replacement for the comparison view, but instead
the motivation is for making the individual test runs easier to
understand.