Make the perfherder graph dynamically sized
This patch makes it so that the graph does not resize the text of the
graph when resizing the window. Previously, the graph had a fixed width,
and the svg would resize with the browser. This created a usability
issue because the graph text and tooltips would become small and
unusable.

This patch adds the WithSize component from the profiler, which provides
the width and the height of the component through the use of a higher
order component.