Fix the positioning of the tooltip for the new dynamic width
The perfherder tooltip relied on some hard coded numbers for the sizing
of the content. This patch makes the sizing more dynamic, so that the
tooltip is correctly positioned at the extremes of the graph.

In addition, it fixes a small issue where the transform property on the
"tip" of the tooltip caused a slight compositing issue where the tip and
the box had a row of pixels that weren't drawn.

It changes the positioning strategy of the box to grow up from
the point, so that the tooltip is always positioned above the point.

It also changes the hit box of the tooltip to be much smaller. The area
under the tooltip "tip" was blocking the hovering behavior of the data
point, making the interaction of the data feel a bit odd.